Grid testing proof of concept
The grid drawing function involves enough code that it should be its own
file. We attempted writing clojurescript tests first, but we can make
the grid file compile for just clojure if we pass in window height and
width to the function, which we should be doing anyways.

But, since setting up the clojurescript tests wasn't quite trivial, this
commit exists as a proof of concept showing how to do it if the grid
namespace was cljs. I'll be removing the cljs versions after this commit
since we don't actually want them, but we may want something like this
in the future.

The "testing" part of the boot invocation is like a "with-profile
test" in lein, although I'm a fan of just constructing a custom task to
do that, I have no reason to run the test tasks without adding the test/
directory to source paths. I do wonder if there is a way to get the test
runner to ignore src/ files, since they won't contain tests, but I
assume removing src/ from source-paths will break things. It's not a big
deal right now though, just results in extra output.

Co-authored by: lindboe <none>