Add a cljs repl
Not knowing how to do development on a clojurescript project, I
investigated how to use a clojurescript repl with boot. Added a task
to do it to the build.boot file and instructions on how to use it to the
README. I wish it were simpler, but this does seem to be the prescribed
workflow.

This is also making me more sure that we should put whatever code we can
into cljc files, to make them useable from typical clojure tooling,
it'll be more comfortable to work with I think.

Also removed the nightlight call in the run task since we don't use it.