A sparse representation of the grid
In this, a grid is a map with a height, width, and state map, where the state map has keys that correspond to coordinates. If those coordinates are occupied, a keyword representing the occupying entity is present in the map. the state map can be iterated over for rendering. still need to experiment with rendering and movement to decide which i like better; and after I do that I'd like to make a system for applying a series of validation tests and returning the tests which were failed, for logging or business logic decisions afteward (many invalid moves are likely to just be ignored, or we can make a sound if trying to move into an occupied space